modify int64 KI->KJ

+++ tbls/market.go
@@ -218,8 +218,8 @@ type Market3 struct {
 type Kline struct{
 	Sym string
 	Time time.Time
-	Open int32
-	High int32
-	Low int32
-	Close int32
+	Open int64
+	High int64
+	Low int64
+	Close int64
 }
